Tighten scope on default scopes.
I am not a fan of these default ORDER BYs; I would much rather that they
not exist at all.  However, unlike the Response default order, it's
quite possible that Surveyor *view code* depends on these orders being
present.  (The same problem exists with any other client of these
models, such as other people's applications.)

If that ends up being true, it's bad software design: a view should
declare its order dependency when it reads data, because that is far
easier to verify as correct.  Unfortunately, like many design issues, it
is not likely to be something that is necessarily trivial to fix.  The
real fix is of course the more appealing one; however, in the battle
between architectural sense and schedule pressures, the former usually
loses.  Hence this hack.